requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
kafka-python>=2.0.2
